Structure the UI in terms of 4 badge lines that you can edit
❤️ Thanks for creating this site - it made it easy for me to spruce up my Universe badge for future GitHub events ✨ ❤️

---

While it's fantastic to have a web UI for updating your badge, with the current implementation, I found it a bit hard, because there are fields that do nothing (e.g. pronouns) and the ordering of them isn't totally logical.

In this PR, I propose a change: switching from labelled fields to "Line 1", "Line 2", "Line 3", etc.

As part of this, I also switch populating a badge from your GitHub profile to an explicit action with a button.

On the side, in the first commit, I have also dropped some unused code related to the QR code and download functionality, which is commented out.

No pressure to accept this in you prefer what you have now, but it was an improvement for me 😅